pahole: Store the class id in 'struct structure' as well
Needed to defer calling printing classes to after we have all sorted out
by name with the upcoming 'pahole --sort' option, needed to make it
possible to compare 'pahole -F btf' with 'pahole -F dwarf -j', as the
multithreaded DWARF loader will not have all classes in a deterministic
order. This is needed for 'btfdiff'.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>