pahole: Fix races in accessing type information in live CUs
When using multithreaded DWARF loading we can't really freely access
some tables, as they may grow and lead to stale data accesses generating
segfaults.

So use a type comparision that takes into account just the immutable
information for structs, unions.

This isn't enough to discern if two types with the same name are really
the same, as we need to look at the member types to figure that out.

So if there are types for which member types need to be checked, leave
it for when all CUs were processed and are thus completely immutable to
resort and fully compare such types.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>