core: Make the tags tables a separate object
And allocate it in cu__new(), add a destructor, etc.

To support DWARF alternate files (.dwz) we'll load the alternate file
passing the tables for the DWARF file being processed so that we can
append its types there while keeping a separate DWARF offset hashtable,
as the same offset may point to different types in the main DWARF file
and in its alternate one.

We will also need to mark on the tag type if it is an alternate one or
not, but that is already available in DWARF, i.e. like this:

  $ readelf -wi /usr/lib/debug/usr/lib64/libvtep-2.17.so.0.0.0-2.17.0-3.fc36.x86_64.debug

   <2><1caf>: Abbrev Number: 33 (DW_TAG_inlined_subroutine)
      <1cb0>   DW_AT_abstract_origin: <alt 0x6a80>
      <1cb4>   DW_AT_entry_pc    : 0x1b804
      <1cbc>   DW_AT_GNU_entry_view: 1
      <1cbd>   DW_AT_ranges      : 0x1d4
      <1cc1>   DW_AT_call_file   : 1
      <1cc2>   DW_AT_call_line   : 251
      <1cc3>   DW_AT_call_column : 5

Then, in its alternate file:

  $ rpm -qf /usr/lib/debug/usr/lib64/libvtep-2.17.so.0.0.0-2.17.0-3.fc36.x86_64.debug
  openvswitch-debuginfo-2.17.0-3.fc36.x86_64
  $ rpm -ql openvswitch-debuginfo-2.17.0-3.fc36.x86_64 | grep dwz
  /usr/lib/debug/.dwz
  /usr/lib/debug/.dwz/openvswitch-2.17.0-3.fc36.x86_64
  $

  $ readelf -wi /usr/lib/debug/.dwz/openvswitch-2.17.0-3.fc36.x86_64 | grep '<6a80>' -A10
   <1><6a80>: Abbrev Number: 22 (DW_TAG_subprogram)
      <6a81>   DW_AT_external    : 1
      <6a81>   DW_AT_name        : (indirect string, offset: 0x1ceab): memset
      <6a85>   DW_AT_decl_file   : 90
      <6a86>   DW_AT_decl_line   : 57
      <6a87>   DW_AT_decl_column : 1
      <6a87>   DW_AT_prototyped  : 1
      <6a87>   DW_AT_type        : <0x1ed2e>
   <2><6a8b>: Abbrev Number: 23 (DW_TAG_formal_parameter)
      <6a8c>   DW_AT_name        : (indirect string, offset: 0x1eb26): __dest
      <6a90>   DW_AT_decl_file   : 90
  $

So that 6a80 DWARF offset may also exist for a given DIE in
/usr/lib/debug/usr/lib64/libvtep-2.17.so.0.0.0-2.17.0-3.fc36.x86_64.debug,
where it would appear as <0x6a80>, not as <alt 0x6a80>.

There is just one alternate per DWARF file, and for
/usr/lib/debug/usr/lib64/libvtep-2.17.so.0.0.0-2.17.0-3.fc36.x86_64.debug
is /usr/lib/debug/.dwz/openvswitch-2.17.0-3.fc36.x86_64.

There will be two DWARF offset keyed hash tables, one for the main file
(libvtep) and another for the alternate (.dwz/openvswitch), but just one
'struct tags_table' instance, combining the types in both.

When resolving the types for the tags from the DWARF offset to the
small_id used in the tags_tables tables we'll do the hash lookups on the
two tables, based in the presence of the 'alt' bit in the type for each
tag.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>