btfdiff: Parallelize loading BTF and DWARF
DWARF is already loaded in parallel, one thread per CPU, we did it for
BTF encoding that is used when building the kernel.

Now just start that in parallel for pretty printing from DWARF and then
do it for BTF, then wait for DWARF to finish to continue and do the
diff.

Not a big deal with a 28 core i7 14700 Intel Rocket Lake refresh with 28
cores (a mix of efficiency and performance cores), but maybe with more
and more cores this may end up making _some_ difference.

The patch is so small, and since I tried it, keep it.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>