btfdiff: Do type expansion to cover "private" types and enumerations
Enumerations are note being printed together with unions and structs,
but they are expanded with 'pahole -E', so using type expansion allows
us to test those as well.

And also helps checking if type expansion from BTF matches the one done
from DWARF. The additional time it takes is negligible and this is is
after all a test script.

Without -E:

  $ time btfdiff vmlinux

  real	0m6.222s
  user	0m13.129s
  sys	0m2.601s
  $

With:

  $ time btfdiff vmlinux

  real	0m6.306s
  user	0m13.230s
  sys	0m2.720s
  $

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>