tests/reproducible_build: Use --btf_features=all when encoding
Alan suggested:

"What about adding --btf_features=all as this would mean we'd be
encoding all the standard kernel features? we'd need to do the same
below in the thread loop. without that we're missing out on a few
features that the kernel builds use in BTF encoding, and we probably
want to ensure that we're testing as close to a real kernel BTF encoding
scenario as possible."

Do it.

Suggested-by: Alan Maguire <alan.maguire@oracle.com>
Cc: Jiri Olsa <jolsa@kernel.org>
Cc: Kui-Feng Lee <kuifeng@fb.com>
Cc: Thomas Weißschuh <linux@weissschuh.net>
Link: https://lore.kernel.org/lkml/3817fa6d-122f-4cbc-92be-616355ec04c2@oracle.com
Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>