pahole: Allow --btf_features to not participate in "all"
Specifying --btf_features=all enables all supported BTF features.

However there are some features that are non-standard, so we should
support a way to use them in --btf_features but not participate in the
set of features enabled by "--btf_features=all".  As part of this, also
support all used in a list of --btf_features, i.e.

  --btf_features=all,nonstandard_feature

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>
Tested-by: Arnaldo Carvalho de Melo <acme@redhat.com>
Cc: Jiri Olsa <jolsa@kernel.org>
Cc: Kui-Feng Lee <kuifeng@fb.com>
Cc: Thomas Weißschuh <linux@weissschuh.net>
Link: https://lore.kernel.org/r/20240416143718.2857981-2-alan.maguire@oracle.com
Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>