core: Introduce per tag alloc method
To allow for the core bits to allocate what is expected by each type
information format, i.e. simple types like CTF and BTF don't need it and
default to zalloc/free, but the DWARF loader needs to have a 'struct
dwarf_tag' associated with each 'struct tag', so allow for a an
allocator that somehow allocates that when a 'struct tag' needs to be
allocated.

Normally this is all done by the type-specific code, but sometimes, like
when doing a class__clone, we better let the type-specific code that
we're creating tags.

Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>